+++ backend/windmill-worker/src/worker.rs
@@ -6,6 +6,7 @@
  * LICENSE-AGPL for a copy of the license.
  */
 
+use const_format::concatcp;
 use itertools::Itertools;
 use sqlx::{Pool, Postgres, Transaction};
 use std::{borrow::Borrow, collections::HashMap, io, panic, process::Stdio, time::Duration};
@@ -35,7 +36,7 @@ use tokio::{
 
 use futures::{
     future::{self, ready, FutureExt},
-    stream::{self, StreamExt},
+    stream, StreamExt,
 };
 
 use async_recursion::async_recursion;
@@ -46,6 +47,59 @@ use crate::{
         handle_flow, update_flow_status_after_job_completion, update_flow_status_in_progress,
     },
 };
+
+#[cfg(feature = "enterprise")]
+async fn run_periodic_jobs(bucket: &str) {
+    tracing::info!("Running periodic jobs");
+
+    match Command::new("rclone")
+        .arg("bisync")
+        .arg(format!(":s3,env_auth=true:{}", bucket))
+        .arg(ROOT_CACHE_DIR)
+        .arg("--size-only")
+        .arg("--fast-list")
+        .arg("--force")
+        .stdin(Stdio::null())
+        .stdout(Stdio::null())
+        .spawn()
+    {
+        Ok(mut h) => {
+            h.wait().await.unwrap();
+        }
+        Err(e) => tracing::warn!("Failed to run periodic job. Error: {:?}", e),
+    }
+}
+
+#[cfg(feature = "enterprise")]
+async fn run_periodic_jobs_initial(bucket: &str) {
+    tracing::info!("Running periodic jobs");
+
+    match Command::new("rclone")
+        .arg("bisync")
+        .arg(format!(":s3,env_auth=true:{}", bucket))
+        .arg(ROOT_CACHE_DIR)
+        .arg("--size-only")
+        .arg("--fast-list")
+        .arg("--force")
+        .arg("--resync")
+        .stdin(Stdio::null())
+        .stdout(Stdio::null())
+        .spawn()
+    {
+        Ok(mut h) => {
+            h.wait().await.unwrap();
+        }
+        Err(e) => tracing::warn!("Failed to run periodic job. Error: {:?}", e),
+    }
+
+    for x in [PIP_CACHE_DIR, DENO_CACHE_DIR, GO_CACHE_DIR] {
+        DirBuilder::new()
+            .recursive(true)
+            .create(x)
+            .await
+            .expect("could not create initial worker dir");
+    }
+}
 
 #[tracing::instrument(level = "trace", skip_all)]
 pub async fn create_token_for_owner<'c>(
@@ -109,31 +163,12 @@ pub async fn create_token_for_owner<'c>(
 }
 
 const TMP_DIR: &str = "/tmp/windmill";
-const PIP_SUPERCACHE_DIR: &str = "/tmp/windmill/cache/pip_permanent";
-const PIP_CACHE_DIR: &str = "/tmp/windmill/cache/pip";
-const DENO_CACHE_DIR: &str = "/tmp/windmill/cache/deno";
-const GO_CACHE_DIR: &str = "/tmp/windmill/cache/go";
+const ROOT_CACHE_DIR: &str = "/tmp/windmill/cache/";
+const PIP_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "pip");
+const DENO_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "deno");
+const GO_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "go");
 const NUM_SECS_ENV_CHECK: u64 = 15;
-const DEFAULT_HEAVY_DEPS: [&str; 18] = [
-    "numpy",
-    "pandas",
-    "anyio",
-    "attrs",
-    "certifi",
-    "h11",
-    "httpcore",
-    "httpx",
-    "idna",
-    "python-dateutil",
-    "rfc3986",
-    "six",
-    "sniffio",
-    "windmill-api",
-    "wmill",
-    "psycopg2-binary",
-    "matplotlib",
-    "seaborn",
-];
+const NUM_SECS_SYNC: u64 = 60 * 10;
 
 const INCLUDE_DEPS_PY_SH_CONTENT: &str = include_str!("../nsjail/download_deps.py.sh");
 const NSJAIL_CONFIG_DOWNLOAD_PY_CONTENT: &str = include_str!("../nsjail/download.py.config.proto");
@@ -192,6 +227,7 @@ pub async fn run_worker(
     ip: &str,
     sleep_queue: u64,
     worker_config: WorkerConfig,
+    sync_bucket: Option<String>,
     mut rx: tokio::sync::broadcast::Receiver<()>,
 ) {
     let start_time = Instant::now();
@@ -199,13 +235,7 @@ pub async fn run_worker(
     let worker_dir = format!("{TMP_DIR}/{worker_name}");
     tracing::debug!(worker_dir = %worker_dir, worker_name = %worker_name, "Creating worker dir");
 
-    for x in [
-        &worker_dir,
-        PIP_SUPERCACHE_DIR,
-        PIP_CACHE_DIR,
-        DENO_CACHE_DIR,
-        GO_CACHE_DIR,
-    ] {
+    for x in [&worker_dir, PIP_CACHE_DIR, DENO_CACHE_DIR, GO_CACHE_DIR] {
         DirBuilder::new()
             .recursive(true)
             .create(x)
@@ -220,7 +250,10 @@ pub async fn run_worker(
     )
     .await;
 
-    let mut last_ping = Instant::now() - Duration::from_secs(NUM_SECS_ENV_CHECK + 1);
+    #[cfg(feature = "enterprise")]
+    let mut last_sync = Instant::now() - Duration::from_secs(NUM_SECS_ENV_CHECK + 1);
+
+    let mut last_ping = Instant::now() - Duration::from_secs(NUM_SECS_SYNC + 1);
 
     insert_initial_ping(worker_instance, &worker_name, ip, db).await;
 
@@ -264,9 +297,6 @@ pub async fn run_worker(
     let go_path = std::env::var("GO_PATH").unwrap_or_else(|_| "/usr/bin/go".to_string());
     let python_path =
         std::env::var("PYTHON_PATH").unwrap_or_else(|_| "/usr/local/bin/python3".to_string());
-    let python_heavy_deps = std::env::var("PYTHON_HEAVY_DEPS")
-        .map(|x| x.split(',').map(|x| x.to_string()).collect::<Vec<_>>())
-        .unwrap_or_else(|_| vec![]);
     let nsjail_path = std::env::var("NSJAIL_PATH").unwrap_or_else(|_| "nsjail".to_string());
     let path_env = std::env::var("PATH").unwrap_or_else(|_| String::new());
     let home_env = std::env::var("HOME").unwrap_or_else(|_| String::new());
@@ -277,7 +307,6 @@ pub async fn run_worker(
         deno_path,
         go_path,
         python_path,
-        python_heavy_deps,
         nsjail_path,
         path_env,
         home_env,
@@ -286,6 +315,11 @@ pub async fn run_worker(
         pip_trusted_host,
     };
     WORKER_STARTED.inc();
+
+    #[cfg(feature = "enterprise")]
+    if let Some(ref s) = sync_bucket.clone() {
+        run_periodic_jobs_initial(&s).await;
+    }
 
     let (same_worker_tx, mut same_worker_rx) = mpsc::channel::<Uuid>(5);
 
@@ -307,6 +341,14 @@ pub async fn run_worker(
                 .expect("update worker ping");
 
                 last_ping = Instant::now();
+            }
+
+            #[cfg(feature = "enterprise")]
+            if last_sync.elapsed().as_secs() > NUM_SECS_SYNC {
+                if let Some(ref s) = sync_bucket.clone() {
+                    run_periodic_jobs(&s).await;
+                }
+                last_sync = Instant::now();
             }
 
             let (do_break, next_job) = async {
@@ -548,7 +590,6 @@ struct Envs {
     deno_path: String,
     go_path: String,
     python_path: String,
-    python_heavy_deps: Vec<String>,
     nsjail_path: String,
     path_env: String,
     home_env: String,
@@ -1370,7 +1411,6 @@ async fn handle_python_job(
     envs @ Envs {
         nsjail_path,
         python_path,
-        python_heavy_deps,
         path_env,
         pip_extra_index_url,
         pip_index_url,
@@ -1416,25 +1456,12 @@ async fn handle_python_job(
                 job_dir,
                 "download.config.proto",
                 &NSJAIL_CONFIG_DOWNLOAD_PY_CONTENT
-                    .replace("{JOB_DIR}", job_dir)
                     .replace("{WORKER_DIR}", &worker_dir)
                     .replace("{CACHE_DIR}", PIP_CACHE_DIR)
                     .replace("{CLONE_NEWUSER}", &(!disable_nuser).to_string()),
             )
             .await?;
         }
-
-        let mut heavy_deps = DEFAULT_HEAVY_DEPS
-            .iter()
-            .map(|s| s.to_string())
-            .collect::<Vec<String>>();
-        heavy_deps.extend(python_heavy_deps.into_iter().map(|s| s.to_string()));
-
-        let (heavy, regular): (Vec<&str>, Vec<&str>) = requirements
-            .split("\n")
-            .partition(|d| heavy_deps.iter().any(|hd| d.starts_with(hd)));
-
-        let _ = write_file(job_dir, "requirements.txt", &regular.join("\n")).await?;
 
         let mut vars = vec![("PATH", path_env)];
         if let Some(url) = pip_extra_index_url {
@@ -1447,80 +1474,20 @@ async fn handle_python_job(
             vars.push(("TRUSTED_HOST", host));
         }
 
-        if heavy.len() > 0 {
-            logs.push_str(&format!(
-                "\nheavy deps detected, using supercache for: {heavy:?}"
-            ));
-            additional_python_paths =
-                handle_python_heavy_reqs(python_path, heavy, vars.clone(), job, logs, db, timeout)
-                    .await?;
-        }
-
-        if regular.len() > 0 {
-            tracing::info!(
-                worker_name = %worker_name,
-                job_id = %job.id,
-                workspace_id = %job.workspace_id,
-                "started setup python dependencies"
-            );
-
-            let child = if !disable_nsjail {
-                Command::new(nsjail_path)
-                    .current_dir(job_dir)
-                    .env_clear()
-                    .envs(vars)
-                    .args(vec!["--config", "download.config.proto"])
-                    .stdout(Stdio::piped())
-                    .stderr(Stdio::piped())
-                    .spawn()?
-            } else {
-                let mut args = vec![
-                    "-m",
-                    "pip",
-                    "install",
-                    "--no-deps",
-                    "--no-color",
-                    "--isolated",
-                    "--no-warn-conflicts",
-                    "--disable-pip-version-check",
-                    "-t",
-                    "./dependencies",
-                    "-r",
-                    "./requirements.txt",
-                ];
-                if let Some(url) = pip_extra_index_url {
-                    args.extend(["--extra-index-url", url]);
-                }
-                if let Some(url) = pip_index_url {
-                    args.extend(["--index-url", url]);
-                }
-                if let Some(host) = pip_trusted_host {
-                    args.extend(["--trusted-host", host]);
-                }
-                Command::new(python_path)
-                    .current_dir(job_dir)
-                    .env_clear()
-                    .envs(vars)
-                    .args(args)
-                    .stdout(Stdio::piped())
-                    .stderr(Stdio::piped())
-                    .spawn()?
-            };
-
-            logs.push_str("\n--- PIP DEPENDENCIES INSTALL ---\n");
-            let child = handle_child(&job.id, db, logs, timeout, child).await;
-            tracing::info!(
-                worker_name = %worker_name,
-                job_id = %job.id,
-                workspace_id = %job.workspace_id,
-                is_ok = child.is_ok(),
-                "finished setting up python dependencies {}",
-                job.id
-            );
-            child?;
-        } else {
-            logs.push_str("\nskipping pip install since not needed");
-        };
+        additional_python_paths = handle_python_reqs(
+            python_path,
+            requirements.split("\n").collect(),
+            vars.clone(),
+            job,
+            logs,
+            db,
+            timeout,
+            nsjail_path,
+            disable_nsjail.clone(),
+            worker_name,
+            job_dir,
+        )
+        .await?;
     }
     logs.push_str("\n\n--- PYTHON CODE EXECUTION ---\n");
 
@@ -1577,10 +1544,7 @@ with open("result.json", 'w') as f:
     write_file(job_dir, "main.py", &wrapper_content).await?;
 
     let mut reserved_variables = get_reserved_variables(job, &token, &base_url, db).await?;
-    let additional_python_paths_folders = additional_python_paths
-        .iter()
-        .map(|x| format!(":{x}"))
-        .join("");
+    let additional_python_paths_folders = additional_python_paths.iter().join(":");
     if !disable_nsjail {
         let shared_deps = additional_python_paths
             .into_iter()
@@ -1612,10 +1576,7 @@ mount {{
         )
         .await?;
     } else {
-        reserved_variables.insert(
-            "PYTHONPATH".to_string(),
-            format!("{job_dir}/dependencies{additional_python_paths_folders}"),
-        );
+        reserved_variables.insert("PYTHONPATH".to_string(), additional_python_paths_folders);
     }
 
     tracing::info!(
@@ -1907,10 +1868,10 @@ async fn pip_compile(
     db: &Pool<Postgres>,
     timeout: i32,
 ) -> error::Result<String> {
-    logs.push_str(&format!("content of requirements:\n{}\n", requirements));
+    logs.push_str(&format!("\ncontent of requirements:\n{}", requirements));
     let file = "requirements.in";
     write_file(job_dir, file, &requirements).await?;
-    let mut args = vec!["-q", "--no-header", file];
+    let mut args = vec!["-q", "--no-header", file, "--resolver=backtracking"];
     if let Some(url) = pip_extra_index_url {
         args.extend(["--extra-index-url", url]);
     }
@@ -2472,50 +2433,84 @@ async fn handle_zombie_jobs(db: &Pool<Postgres>, timeout: i32, base_url: &str) {
     }
 }
 
-async fn handle_python_heavy_reqs(
+async fn handle_python_reqs(
     python_path: &String,
-    heavy_requirements: Vec<&str>,
+    requirements: Vec<&str>,
     vars: Vec<(&str, &String)>,
     job: &QueuedJob,
     logs: &mut String,
     db: &sqlx::Pool<sqlx::Postgres>,
     timeout: i32,
+    nsjail_path: &str,
+    disable_nsjail: bool,
+    worker_name: &str,
+    job_dir: &str,
 ) -> error::Result<Vec<String>> {
     let mut req_paths: Vec<String> = vec![];
-    for req in heavy_requirements {
+    for req in requirements {
         // todo: handle many reqs
-        let venv_p = format!("{PIP_SUPERCACHE_DIR}/{req}");
+        let venv_p = format!("{PIP_CACHE_DIR}/{req}");
         if metadata(&venv_p).await.is_ok() {
-            tracing::info!("already exists: {:?}", &venv_p);
             req_paths.push(venv_p);
             continue;
         }
 
-        logs.push_str("\n--- PIP SUPERCACHE INSTALL ---\n");
-        logs.push_str(&format!("\nthe heavy dependency {req} is being installed for the first time.\nIt will take a bit longer but further execution will be much faster!"));
+        logs.push_str("\n--- PIP INSTALL ---\n");
+        logs.push_str(&format!("\n{req} is being installed for the first time.\n It will be cached for all ulterior uses."));
 
-        logs.push_str("pip install\n");
-        let child = Command::new(python_path)
-            .env_clear()
-            .envs(vars.clone())
-            .args(vec![
-                "-m",
-                "pip",
-                "install",
-                &req,
-                "-I",
-                "--no-deps",
-                "--no-color",
-                "--isolated",
-                "--no-warn-conflicts",
-                "--disable-pip-version-check",
-                "-t",
-                venv_p.as_str(),
-            ])
-            .stdout(Stdio::piped())
-            .stderr(Stdio::piped())
-            .spawn()?;
-        handle_child(&job.id, db, logs, timeout, child).await?;
+        tracing::info!(
+            worker_name = %worker_name,
+            job_id = %job.id,
+            workspace_id = %job.workspace_id,
+            "started setup python dependencies"
+        );
+
+        let child = if !disable_nsjail {
+            let mut vars = vars.clone();
+            let req = req.to_string();
+            vars.push(("REQ", &req));
+            vars.push(("TARGET", &venv_p));
+            Command::new(nsjail_path)
+                .current_dir(job_dir)
+                .env_clear()
+                .envs(vars)
+                .args(vec!["--config", "download.config.proto"])
+                .stdout(Stdio::piped())
+                .stderr(Stdio::piped())
+                .spawn()?
+        } else {
+            Command::new(python_path)
+                .env_clear()
+                .envs(vars.clone())
+                .args(vec![
+                    "-m",
+                    "pip",
+                    "install",
+                    &req,
+                    "-I",
+                    "--no-deps",
+                    "--no-color",
+                    "--isolated",
+                    "--no-warn-conflicts",
+                    "--disable-pip-version-check",
+                    "-t",
+                    venv_p.as_str(),
+                ])
+                .stdout(Stdio::piped())
+                .stderr(Stdio::piped())
+                .spawn()?
+        };
+
+        let child = handle_child(&job.id, db, logs, timeout, child).await;
+        tracing::info!(
+            worker_name = %worker_name,
+            job_id = %job.id,
+            workspace_id = %job.workspace_id,
+            is_ok = child.is_ok(),
+            "finished setting up python dependencies {}",
+            job.id
+        );
+        child?;
 
         req_paths.push(venv_p);
     }