+++ frontend/src/lib/components/apps/editor/ComponentEditor.svelte
@@ -8,6 +8,7 @@
 	import TextComponent from '../components/common/TextComponent.svelte'
 	import type { AppComponent, AppEditorContext } from '../types'
 	import { displayData } from '../utils'
+	import ButtonComponent from '../components/common/ButtonComponent.svelte'
 
 	export let component: AppComponent
 	export let selected: boolean
@@ -45,9 +46,19 @@
 		{:else if component.type === 'barchartcomponent'}
 			<BarChartComponent {...component} bind:staticOutputs={$staticOutputs[component.id]} />
 		{:else if component.type === 'tablecomponent'}
-			<TableComponent {...component} bind:staticOutputs={$staticOutputs[component.id]} />
+			<TableComponent
+				{...component}
+				bind:staticOutputs={$staticOutputs[component.id]}
+				bind:inputs={component.inputs}
+			/>
 		{:else if component.type === 'textcomponent'}
 			<TextComponent
+				{...component}
+				bind:staticOutputs={$staticOutputs[component.id]}
+				bind:componentInputs={component.componentInputs}
+			/>
+		{:else if component.type === 'buttoncomponent'}
+			<ButtonComponent
 				{...component}
 				bind:staticOutputs={$staticOutputs[component.id]}
 				bind:componentInputs={component.componentInputs}